requirements.txt
# AI Employee - Python Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# CORE DEPENDENCIES (Required)
# =============================================================================

# Async support
asyncio-throttle>=1.0.2
aiohttp>=3.9.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.0

# Date/Time handling
python-dateutil>=2.8.2
croniter>=2.0.0

# HTTP/API
httpx>=0.25.0

# CLI tools
click>=8.1.0
rich>=13.7.0

# Logging
structlog>=23.2.0

# =============================================================================
# GOOGLE INTEGRATIONS (Gmail, Calendar)
# =============================================================================

google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0

# =============================================================================
# COMMUNICATION INTEGRATIONS
# =============================================================================

# Slack
slack-sdk>=3.23.0

# Discord
discord.py>=2.3.0

# WhatsApp (via Node.js bridge - see docs)
# Requires: npm install whatsapp-web.js qrcode-terminal

# Twitter/X
tweepy>=4.14.0

# LinkedIn (unofficial API)
linkedin-api>=2.0.0

# SMS (Twilio)
twilio>=8.10.0

# =============================================================================
# FILE HANDLING
# =============================================================================

watchdog>=3.0.0
python-magic>=0.4.27
PyPDF2>=3.0.0
Pillow>=10.0.0

# =============================================================================
# TESTING
# =============================================================================

pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# =============================================================================
# DEVELOPMENT (Optional)
# =============================================================================

black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# =============================================================================
# ADDITIONAL UTILITIES
# =============================================================================

# For invoice/document processing
pytesseract>=0.3.10
pdf2image>=1.16.0

# For data export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# For charts/visualization (optional)
matplotlib>=3.8.0

# Encryption for sensitive data
cryptography>=41.0.0
